requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
pillow>=10.0.0
pypdf2==3.0.1
pdf2image==1.17.0
python-dotenv==1.0.0
openai==1.3.7
google-generativeai==0.3.2
aiofiles==23.2.1
requests==2.31.0
numpy>=1.24.0
opencv-python-headless==4.8.1.78
typing-extensions>=4.8.0
pydantic-settings==2.1.0